pyproject.toml
[tool.poetry]
name = "rent4me"
version = "0.1.0"
description = ""
authors = ["Daniil <39118675+TyPaporotnyk@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
psycopg2 = "^2.9.9"
django-environ = "^0.11.2"
django-ninja = "^1.2.0"
punq = "^0.7.0"
orjson = "^3.10.6"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"
black = "^24.4.2"
isort = "^5.13.2"
pytest = "^8.2.2"
pytest-django = "^4.8.0"
pre-commit = "^3.7.1"
faker = "^26.0.0"
factory-boy = "^3.3.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.local"

[tool.black]
line-length = 119
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 119
skip = ["migrations/"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
